# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "In digital",
        "description": "This is a microservice, which provide many services as part of InDigital Challenge.",
        "contact": {
            "name": "Gian Huachin",
            "email": "ghuachins@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://in-digital-challenge.local.vh:8108",
            "description": "development environment"
        }
    ],
    "paths": {
        "/creacliente": {
            "post": {
                "tags": [
                    "client"
                ],
                "description": "Crear cliente",
                "operationId": "createClient",
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "description": "Nombre del cliente",
                                        "type": "string"
                                    },
                                    "apellido": {
                                        "description": "Apellido del cliente",
                                        "type": "string"
                                    },
                                    "fecha_nacimiento": {
                                        "description": "Fecha de nacimiento del cliente",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cliente creado exitosamente"
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                }
            }
        },
...
    },
...
}